Update to 2.0.1

+++ src/commands/get_help.py
@@ -7,26 +7,23 @@ This file can also be imported as a module and contains the following functions:
 """
 
 
+from discord import channel
 import asyncio
 
 
 HELP_MESSAGE = 'Доступные команды бота: ' \
                '\n\n**полл (время и текст | текст)** - запускает простое голосование' \
                '\n**макар** - возвращает предложение "Улыбок тебе ' \
-               'дед [перевёрнутое предложение]"' \
+               'дед [перевёрнутая строка]"' \
                '\n**хелп** - выводит информацию с командами' \
-               '\n**uptime** - выводит сколько бот проработал ' \
-               'с последнего запуска на сервере' \
+               '\n**uptime** - выводит время работы бота' \
                '\n**йа (анон | анонттс)** - аналог команды `/me`' \
                '\n**рандом (число | два числа)** - получение рандомного числа' \
-               '\n**шип (два имени)** - шипперит двух рандомных пользователей и ' \
-               'скрепляет их с помощью спаренного имени' \
+               '\n**шип (два имени)** - шипперит двух рандомных пользователей' \
                '\n**цуефа** - игра в "Камень Ножницы Бумага"' \
                '\n**рулетка** - запускает игру в русскую рулетку' \
-               '\n**система** - показывает данные о системе, ' \
-               'на которой запущен бот' \
-               '\n**ху | who** - выбирает рандомного пользователя и показывает ему ' \
-               'рандомное предложение из массива слов' \
+               '\n**система** - показывает данные о системе' \
+               '\n**ху | who** - рандомный пользователь + рандомное предложение' \
                '\n**поиск (пидорасов)** -  *поиск пидорасов активирован...*'
 DELAY_TIME = 5
 
@@ -37,9 +34,12 @@ async def send_help_message(msg):
     Parameters:
         msg (discord.message.Message): Execute send to channel function
     """
-    await msg.author.send(HELP_MESSAGE)
-    warn_message = await msg.channel.send(f'{msg.author.mention}, проверь личку! '
-                                          'Я отправил тебе помощь по командам',
-                                          delete_after=DELAY_TIME)
-    await asyncio.sleep(DELAY_TIME)
-    await warn_message.delete()
+    if isinstance(msg.channel, channel.DMChannel):
+        await msg.channel.send(HELP_MESSAGE)
+    else:
+        await msg.author.send(HELP_MESSAGE)
+        warn_message = await msg.channel.send(f'{msg.author.mention}, проверь личку! '
+                                              'Я отправил тебе помощь по командам',
+                                              delete_after=DELAY_TIME)
+        await asyncio.sleep(DELAY_TIME)
+        await warn_message.delete()
